almost done

+++ src/edit.js
@@ -29,7 +29,11 @@ function Edit(){
     const deleteNote = () => {
         console.log("delete");
         // remove the note from the list of notes
-        onDelete(current);
+        // add a windo allert checking if they accualt want to delete the note
+        const confirm = window.confirm("Are you sure you want to delete this note?");
+        if(confirm){
+            onDelete(current);
+        }
     }
 
     useEffect(()=> {
@@ -41,16 +45,15 @@ function Edit(){
         <>
             {/* bar to put title in */}
             {current && current.id ? (
-            <div>
-                <h2>{current.title}</h2>
-                <p>{current.date}</p>
-                <div dangerouslySetInnerHTML={{__html: current.contents}}></div>
-                {/* <p>{current.contents}</p> */}
+            <div id="view-box">
+                <h2 id="view-title">{current.title}</h2>
+                <p id="view-date">{current.date}</p>
+                <div id="view-content" dangerouslySetInnerHTML={{__html: current.contents}}></div>
             </div>
             ) : null }
             <div id="note-header-box">
                 <div id="note-title-box">
-                <input id="note-title" type="text" placeholder="Untitled" onChange={(e) => setCurrent((prev) => ({...prev, title: e.target.value}))}></input>
+                <input id="note-title" type="text" placeholder="Untitled" value={current.title} onChange={(e) => setCurrent((prev) => ({...prev, title: e.target.value}))}></input>
                 <input id="note-date" type="datetime-local" onChange={(e) => setCurrent((prev) => ({...prev, date: e.target.value}))}/>
                 </div>
 
@@ -62,8 +65,8 @@ function Edit(){
 
             {/* making note tools */}
             <div id="note-input-container">
-                {/* <textarea id="note-contents" placeholder="Type here..." rows="4" cols="50" onChange={(e) => setCurrent((prev) => ({...prev, contents: e.target.value}))}></textarea> */}
-                <ReactQuill id="note-contents" placeholder="Type here..." onChange={(e) => setCurrent((prev) => ({...prev, contents: e}))}/>
+                {/* <ReactQuill id="note-contents" placeholder="Type here..." onChange={(e) => setCurrent((prev) => ({...prev, contents: e}))}/> */}
+                <ReactQuill id="note-contents" placeholder="Type here..." value={current.contents} onChange={(e) => setCurrent((prev) => ({...prev, contents: e}))}/>
             </div>
         </>
